requirements.txt
Django==1.11.5
django-rest-framework==0.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.3.1
PyHamcrest==1.9.0
pytest==3.5.0
pytest-django==3.1.2
factory-boy==2.10.0
django-mailbox==4.7.1
